qat refactor for nextest
