+++ .cargo/config.toml
@@ -27,12 +27,13 @@ dev-coverage-compare = "run --quiet --package xtask -- coverage compare --lines 
 dev-coverage-html = "llvm-cov --manifest-path bitloops/Cargo.toml --workspace --all-targets --features slow-tests --no-default-features --html --output-dir bitloops/target/llvm-cov-html"
 
 # QAT aliases from workspace root.
-qat = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_acceptance --run-ignored only --no-capture -- qat --exact"
-qat-smoke = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_acceptance --run-ignored only -- qat_smoke --exact"
-qat-devql-capabilities = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_acceptance --run-ignored only -- qat_devql_capabilities --exact"
-qat-devql-sync = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_acceptance --run-ignored only -- qat_devql_sync --exact"
-qat-onboarding = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_acceptance --run-ignored only -- qat_onboarding --exact"
-qat-devql-ingest = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_acceptance --run-ignored only -- qat_devql_ingest --exact"
+qat = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat --run-ignored only -- qat --exact"
+qat-smoke = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_smoke --run-ignored only -- qat_smoke --exact"
+qat-devql-capabilities = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_devql_capabilities --run-ignored only -- qat_devql_capabilities --exact"
+qat-devql-sync = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_devql_sync --run-ignored only -- qat_devql_sync --exact"
+qat-onboarding = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_onboarding --run-ignored only -- qat_onboarding --exact"
+qat-devql-ingest = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_devql_ingest --run-ignored only -- qat_devql_ingest --exact"
+qat-quickstart = "nextest run --manifest-path bitloops/Cargo.toml --features qat-tests --test qat_quickstart --run-ignored only -- qat_quickstart --exact"
 
 # Explicit bundled-DuckDB aliases for release/offline/exotic targets.
 dev-check-bundled = "check --manifest-path bitloops/Cargo.toml --features duckdb-bundled"
