+++ TESTING.md
@@ -36,6 +36,7 @@ For other platforms, follow the official installation guide:
 | DevQL capabilities suite                         | `cargo qat-devql-capabilities` |
 | DevQL sync suite                                 | `cargo qat-devql-sync`       |
 | DevQL ingest suite                               | `cargo qat-devql-ingest`     |
+| Quickstart suite                                 | `cargo qat-quickstart`       |
 
 `cargo dev-loop` runs: `fmt` (write fixes) -> `clippy` -> fast tests -> file-size check.
 `cargo dev-test-fast` is the default local feedback loop.
@@ -43,15 +44,17 @@ For other platforms, follow the official installation guide:
 That default does not ban `cargo test`: use the checked-in aliases for the standard lanes, and use `cargo test` only where this guide explicitly calls for it or where `cargo-nextest` cannot cover the case.
 The checked-in local `nextest` default is `8` test threads.
 CI uses the `ci` `nextest` profile, pinned to `6` test threads.
+The legacy `qat_acceptance` integration target has been split into `qat`, `qat_smoke`, `qat_devql_capabilities`, `qat_devql_ingest`, `qat_devql_sync`, `qat_onboarding`, and `qat_quickstart`.
 `cargo dev-test-merge` runs the fast lane plus a curated set of slow smoke suites and is the blocking gate for pull requests into `develop`.
 `cargo dev-test-slow` runs all slow targets only.
 `cargo dev-test-full` runs fast + slow and is used for post-merge verification on `develop` and pull requests into `main`.
 On macOS, `dev-test-*` and `dev-install` automatically sign produced binaries to reduce repeated policy validation overhead (`syspolicyd`).
 `cargo qat` runs onboarding, smoke, DevQL sync, DevQL capabilities, and DevQL ingest in parallel.
-`cargo qat` forces `--no-capture` so the bundled ignored QAT journey streams progress reliably during long daemon-backed runs.
+`cargo qat` runs the bundled ignored QAT journey through `cargo-nextest`.
 `cargo qat-devql-capabilities` is the focused DevQL capabilities alias.
 `cargo qat-devql-ingest` is the focused DevQL ingest alias.
 `cargo qat-devql-sync` is the focused DevQL sync alias.
+`cargo qat-quickstart` is the focused quickstart alias.
 
 ### QAT scenario filtering
 
@@ -70,7 +73,7 @@ CUCUMBER_FILTER_TAGS='@test_harness_sync' \
 cargo test \
   --manifest-path bitloops/Cargo.toml \
   --features qat-tests \
-  --test qat_acceptance \
+  --test qat_devql_sync \
   qat_devql_sync \
   -- --ignored --nocapture
 ```
