+++ bitloops/tests/qat_internal.rs
@@ -1,0 +1,4 @@
+#![allow(dead_code)]
+
+#[path = "qat_internal_support.rs"]
+mod qat_support;
