+++ bitloops/tests/qat_support/internal_entrypoints.rs
@@ -1,0 +1,5 @@
+include!("entrypoints.rs");
+
+#[cfg(test)]
+#[path = "entrypoints_tests.rs"]
+mod tests;
