+++ bitloops/tests/qat_support/internal_helpers.rs
@@ -1,0 +1,5 @@
+include!("helpers/mod.rs");
+
+#[cfg(test)]
+#[path = "helpers/tests.rs"]
+mod tests;
