+++ bitloops/tests/qat_support/internal_runner.rs
@@ -1,0 +1,5 @@
+include!("runner.rs");
+
+#[cfg(test)]
+#[path = "runner_tests.rs"]
+mod tests;
