+++ bitloops/tests/qat_support/mod.rs
@@ -1,4 +1,5 @@
 pub mod bundle;
+pub mod entrypoints;
 pub mod helpers;
 pub mod runner;
 pub mod steps;
